third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/background_sync/oneshot-register-failure-worker-not-activated-expected.txt

This is a testharness.js-based test.
PASS sync.register() is rejected when the service worker is not active yet 
Harness: the test ran to completion.
